server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
]
